make rarfile dependency optional
